+++ packages/@glimmer/runtime/lib/vm/update.ts
@@ -75,10 +75,6 @@ export default class UpdatingVM {
     this.frame.handleException();
     this.frameStack.pop();
   }
-
-  evaluateOpcode(opcode: UpdatingOpcode) {
-    opcode.evaluate(this);
-  }
 }
 
 export interface ExceptionHandler {
