+++ packages/@glimmer/util/index.ts
@@ -2,10 +2,8 @@ export interface Destroyable {
   destroy(): void;
 }
 
-export { getAttrNamespace } from './lib/namespaces';
 export * from './lib/platform-utils';
 export { default as assert } from './lib/assert';
-export { default as LOGGER, Logger, LogLevel } from './lib/logger';
 
 export { assign, fillNulls } from './lib/object-utils';
 export { ensureGuid, initializeGuid, HasGuid } from './lib/guid';
