+++ packages/@glimmer/util/lib/collections.ts
@@ -8,7 +8,6 @@ export interface Dict<T> {
 export interface Set<T> {
   add(value: T): Set<T>;
   delete(value: T): void;
-  forEach(callback: (item: T) => void): void;
 }
 
 let proto = Object.create(null, {
@@ -51,27 +50,11 @@ export class DictSet<T extends SetMember> implements Set<T> {
     if (typeof obj === 'string') delete this.dict[<any>obj];
     else if ((obj as any)._guid) delete this.dict[(obj as any)._guid];
   }
-
-  forEach(callback: (item: T) => void) {
-    let { dict } = this;
-    let dictKeys = Object.keys(dict);
-    for (let i = 0; dictKeys.length; i++) {
-      callback(dict[dictKeys[i]]);
-    }
-  }
-
-  toArray(): string[] {
-    return Object.keys(this.dict);
-  }
 }
 
 export class Stack<T> {
   private stack: T[] = [];
   public current: Option<T> = null;
-
-  toArray(): T[] {
-    return this.stack;
-  }
 
   push(item: T) {
     this.current = item;
